Dividindo o arquivo para melhor sustentação de código.
