+++ crud.c
@@ -2,227 +2,9 @@
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
-
-// Definindo a estrutura do funcionario
-typedef struct funcionario
-{
-    char cpf[11];
-    char nome[100];
-    char cargo[100];
-    int mes;
-    int ano;
-    float salario;
-    struct funcionario *prox;
-} Funcionario;
-
-typedef struct produto
-{
-    int codigo_produto;
-    char desc_produto[30];
-    char armazem[2];
-    char endereco[5];
-    float custo_unitario;
-    struct produto *prox;
-} Produto;
-
-// Definindo uma estrutura para representar uma lista encadeada
-typedef struct lista
-{
-    Funcionario *inicio;
-    int tamanho;
-} Lista;
-
-// Criando uma função para inicializar uma lista vazia
-Lista *criar_lista()
-{
-    Lista *lista = (Lista *)malloc(sizeof(Lista));
-    lista->inicio = NULL;
-    lista->tamanho = 0;
-    return lista;
-}
-
-// Criando uma função para liberar a memoria de uma lista
-void liberar_lista(Lista *lista)
-{
-    Funcionario *aux = lista->inicio;
-    while (aux != NULL)
-    {
-        Funcionario *prox = aux->prox;
-        free(aux);
-        aux = prox;
-    }
-    free(lista);
-}
-
-// Criando uma função para inserir um funcionario no inicio da lista
-void inserir_inicio(Lista *lista, Funcionario *funcionario)
-{
-    funcionario->prox = lista->inicio;
-    lista->inicio = funcionario;
-    lista->tamanho++;
-}
-
-// Criando uma função para inserir um funcionario no final da lista
-void inserir_final(Lista *lista, Funcionario *funcionario)
-{
-    if (lista->inicio == NULL)
-    {
-        inserir_inicio(lista, funcionario);
-    }
-    else
-    {
-        Funcionario *aux = lista->inicio;
-        while (aux->prox != NULL)
-        {
-            aux = aux->prox;
-        }
-        aux->prox = funcionario;
-        funcionario->prox = NULL;
-        lista->tamanho++;
-    }
-}
-
-// Criando uma função para remover um funcionario da lista pelo cpf
-int remover_funcionario(Lista *lista, char *cpf)
-{
-    if (lista->inicio == NULL)
-    {
-        return 0; // Lista vazia
-    }
-    Funcionario *ant = NULL;
-    Funcionario *aux = lista->inicio;
-    while (aux != NULL && strcmp(aux->cpf, cpf) != 0)
-    {
-        ant = aux;
-        aux = aux->prox;
-    }
-    if (aux == NULL)
-    {
-        return 0; // cpf não encontrado
-    }
-    if (ant == NULL)
-    {
-        lista->inicio = aux->prox;
-    }
-    else
-    {
-        ant->prox = aux->prox;
-    }
-    free(aux);
-    lista->tamanho--;
-    return 1; // cpf removido com sucesso
-}
-
-// Criando uma função para buscar um funcionario na lista pelo cpf
-Funcionario *buscar_funcionario(Lista *lista, char *cpf)
-{
-    Funcionario *aux = lista->inicio;
-    while (aux != NULL && strcmp(aux->cpf, cpf) != 0)
-    {
-        aux = aux->prox;
-    }
-    return aux; // Retorna o funcionario encontrado ou NULL se nÃ£o encontrar
-}
-
-// Criando uma função para imprimir o funcionario na tela
-void imprimir_funcionario(Funcionario *funcionario)
-{
-    printf("Cpf: %s\n", funcionario->cpf);
-    printf("Nome: %s\n", funcionario->nome);
-    printf("Cargo: %s\n", funcionario->cargo);
-    printf("Mes: %d\n", funcionario->mes);
-    printf("Ano: %d\n", funcionario->ano);
-    printf("Salario: %f\n", funcionario->salario);
-}
-
-// Criando uma função para imprimir a lista de funcionarios na tela
-void imprimir_lista(Lista *lista)
-{
-    Funcionario *aux = lista->inicio;
-    while (aux != NULL)
-    {
-        imprimir_funcionario(aux);
-        printf("\n");
-        aux = aux->prox;
-    }
-}
-
-// Criando uma função para ler os dados de um funcionario a partir do teclado
-Funcionario *ler_funcionario()
-{
-    Funcionario *funcionario = (Funcionario *)malloc(sizeof(Funcionario));
-    printf("Digite o cpf do funcionario: ");
-    scanf("%[^\n]", funcionario->cpf);
-    getchar();
-    printf("Digite o nome do funcionario: ");
-    scanf("%[^\n]", funcionario->nome);
-    getchar();
-    printf("Digite o mes de admissao do funcionario: ");
-    scanf(" %d", &funcionario->mes);
-    getchar();
-    printf("Digite o ano de admissao do funcionario: ");
-    scanf(" %d", &funcionario->ano);
-    getchar();
-    printf("Digite o salario do funcionario: ");
-    scanf(" %f", &funcionario->salario);
-    getchar();
-    printf("Digite o cargo do funcionario: ");
-    scanf("%[^\n]", funcionario->cargo);
-    getchar();
-    return funcionario;
-}
-
-// Criando uma função para carregar os dados dos funcionarios a partir de um arquivo
-Lista *carregar_arquivo(char *nome_arquivo)
-{
-    Lista *lista = criar_lista();
-    FILE *arq = fopen(nome_arquivo, "r");
-    if (arq == NULL)
-    {
-        printf("Erro ao abrir o arquivo.\n");
-        return lista;
-    }
-    char linha[300];
-    fgets(linha, 300, arq); // Ignorando a primeira linha do arquivo (cabecalho)
-    while (fgets(linha, 300, arq) != NULL)
-    {
-        Funcionario *funcionario = (Funcionario *)malloc(sizeof(Funcionario));
-        char *token = strtok(linha, ";");
-        strcpy(funcionario->cpf, token);
-        token = strtok(NULL, ";");
-        strcpy(funcionario->nome, token);
-        token = strtok(NULL, ";");
-        funcionario->mes = atoi(token);
-        token = strtok(NULL, ";");
-        funcionario->ano = atoi(token);
-        token = strtok(NULL, ";");
-        funcionario->salario = atof(token);
-        token = strtok(NULL, "\n");
-        strcpy(funcionario->cargo, token);
-        inserir_final(lista, funcionario);
-    }
-    fclose(arq);
-    return lista;
-}
-
-// Criando uma função para salvar os dados dos funcionarios em um arquivo
-void salvar_arquivo(char *nome_arquivo, Lista *lista)
-{
-    FILE *arq = fopen(nome_arquivo, "w");
-    if (arq == NULL)
-    {
-        printf("Erro ao abrir o arquivo.\n");
-        return;
-    }
-    fprintf(arq, "cpf;nome_func;mes;ano;salario;cargo\n"); // Escrevendo o cabeÃ§alho do arquivo
-    Funcionario *aux = lista->inicio;
-    while (aux != NULL)
-    {
-        fprintf(arq, "%s;%s;%d;%d;%f;%s\n", aux->cpf, aux->nome, aux->mes, aux->ano, aux->salario, aux->cargo); // Escrevendo os dados de cada funcionario no arquivo
-        aux = aux->prox;
-    }
-    fclose(arq);
-}
+#include "structdef.h"
+#include "listman.h"
+#include "fileman.h"
 
 void cabecalho()
 {
